refactor setting allowance for exchange services, apply to synthetix
